Refactored Peters script into a generic one.
fix-x.y.z-dependencies now just sources fix-any-dependencies.sh with
a single argument specifying the filename (witout deb-extension9to
fix dependencies on.

So "ibm-notes-9.0.1.i586" can be passed to modify the file
"ibm-notes-9.0.1.i586.deb". The argument also specifies which control-file
to use since the control-files are all named $1_control.

So when a new broken version of IBM/Lotus Notes comes out all that has to be
done is to add the control-file and a script that calls fix-any-dependencies.sh
with the appropriate file-name.